+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [1.0.2] - 2019-02-01
+### Note
+- Added support for websites scope
+- Add compatability for PHP 7.2
+- Fix storeresolver when using multiple websites
+- Add compatibility if storegroup has storeviews all with same url
 
 ## [1.0.0] - 2017-02-03
 ### Note
 - Seperated Ho_Multistore and made StoreResolver into own module
+
+## [Unreleased]